requirements.txt
langchain-community==0.4.1
fastapi-cli==0.0.16
langchain-deepseek==1.0.1
pydantic==2.12.5
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
requests==2.32.3
cors==0.9.11
redis==5.0.1
langchain==1.1.0
langchain-core==1.1.0